Address libdragon n64 code warnings

+++ src/surface.c
@@ -9,6 +9,7 @@
 #include "debug.h"
 #include <assert.h>
 #include <string.h>
+#include <inttypes.h>
 
 const char* tex_format_name(tex_format_t fmt)
 {
@@ -28,7 +29,7 @@ const char* tex_format_name(tex_format_t fmt)
     }
 }
 
-surface_t surface_alloc(tex_format_t format, uint32_t width, uint32_t height)
+surface_t surface_alloc(tex_format_t format, uint16_t width, uint16_t height)
 {
     // A common mistake is to call surface_alloc with the wrong argument order.
     // Try to catch it by checking that the format is not valid.
@@ -53,14 +54,14 @@ void surface_free(surface_t *surface)
     memset(surface, 0, sizeof(surface_t));
 }
 
-surface_t surface_make_sub(surface_t *parent, uint32_t x0, uint32_t y0, uint32_t width, uint32_t height)
+surface_t surface_make_sub(surface_t *parent, uint16_t x0, uint16_t y0, uint16_t width, uint16_t height)
 {
     assert(x0 + width <= parent->width);
     assert(y0 + height <= parent->height);
 
     tex_format_t fmt = surface_get_format(parent);
     assertf(TEX_FORMAT_BITDEPTH(fmt) != 4 || (x0 & 1) == 0,
-        "cannot create a subsurface with an odd X offset (%ld) in a 4bpp surface", x0);
+        "cannot create a subsurface with an odd X offset (%" PRId16 ") in a 4bpp surface", x0);
 
     surface_t sub;
     sub.buffer = parent->buffer + y0 * parent->stride + TEX_FORMAT_PIX2BYTES(fmt, x0);
@@ -71,7 +72,7 @@ surface_t surface_make_sub(surface_t *parent, uint32_t x0, uint32_t y0, uint32_t
     return sub;
 }
 
-extern inline surface_t surface_make(void *buffer, tex_format_t format, uint32_t width, uint32_t height, uint32_t stride);
+extern inline surface_t surface_make(void *buffer, tex_format_t format, uint16_t width, uint16_t height, uint16_t stride);
 extern inline tex_format_t surface_get_format(const surface_t *surface);
-extern inline surface_t surface_make_linear(void *buffer, tex_format_t format, uint32_t width, uint32_t height);
+extern inline surface_t surface_make_linear(void *buffer, tex_format_t format, uint16_t width, uint16_t height);
 extern inline bool surface_has_owned_buffer(const surface_t *surface);
